feat: Support include inside functions (#162)
This patch adds a unified way to call #include when defining functions.

The first three ways send the included module to a wrapper class
inheriting from Callable to keep expected behavior users will likely
expect and be referencing from documentation when calling include.

The fourth and preferred way to call #include is directly on the
function instance with the reason for this being that the include is
immediately applied once instead of each time the function is called.

The main purpose for this approach is to isolate includes to function
definitions. This supports block, callable class, callable objects, and
proc-likes.

Reference the docs/ examples/ and test/test_function.rb for usage

Signed-off-by: Johnneylee Jack Rollins <git@spacegho.st>